+++ visualization/app/codeCharta/state/store.service.ts
@@ -13,6 +13,7 @@ import { IsAttributeSideBarVisibleActions } from "./store/appSettings/isAttribut
 import { PanelSelectionActions } from "./store/appSettings/panelSelection/panelSelection.actions"
 import { PresentationModeActions } from "./store/appSettings/isPresentationMode/isPresentationMode.actions"
 import { ExperimentalFeaturesEnabledActions } from "./store/appSettings/enableExperimentalFeatures/experimentalFeaturesEnabled.actions"
+import { TrackingDataEnabledActions } from "./store/appSettings/enableTrackingData/trackingDataEnabled.actions"
 
 export interface StoreSubscriber {
 	onStoreChanged(actionType: string)
@@ -43,6 +44,7 @@ export class StoreService {
 				isActionOfType(action.type, PanelSelectionActions) ||
 				isActionOfType(action.type, PresentationModeActions) ||
 				isActionOfType(action.type, ExperimentalFeaturesEnabledActions) ||
+				isActionOfType(action.type, TrackingDataEnabledActions) ||
 				options.silent
 			)
 		) {