+++ visualization/app/codeCharta/state/store/appSettings/enableTrackingData/trackingDataEnabled.splitter.ts
@@ -1,0 +1,5 @@
+import { setTrackingDataEnabled } from "./trackingDataEnabled.actions"
+
+export function splitTrackingDataEnabledAction(payload: boolean) {
+	return setTrackingDataEnabled(payload)
+}